+++ sentry-spring-boot-jakarta/src/test/kotlin/io/sentry/spring/boot/jakarta/SentryAutoConfigurationTest.kt
@@ -26,6 +26,8 @@ import io.sentry.spring.jakarta.SentryUserFilter
 import io.sentry.spring.jakarta.SentryUserProvider
 import io.sentry.spring.jakarta.SpringSecuritySentryUserProvider
 import io.sentry.spring.jakarta.tracing.SentryTracingFilter
+import io.sentry.spring.jakarta.tracing.SpringServletTransactionNameProvider
+import io.sentry.spring.jakarta.tracing.TransactionNameProvider
 import io.sentry.transport.ITransport
 import io.sentry.transport.ITransportGate
 import io.sentry.transport.apache.ApacheHttpClientTransportFactory
@@ -442,6 +444,15 @@ class SentryAutoConfigurationTest {
             .withClassLoader(FilteredClassLoader(HandlerExceptionResolver::class.java))
             .run {
                 assertThat(it).doesNotHaveBean(SentryExceptionResolver::class.java)
+            }
+    }
+
+    @Test
+    fun `when Spring MVC is not on the classpath, fallback TransactionNameProvider is configured`() {
+        contextRunner.withPropertyValues("sentry.dsn=http://key@localhost/proj", "sentry.send-default-pii=true")
+            .withClassLoader(FilteredClassLoader(HandlerExceptionResolver::class.java))
+            .run {
+                assertThat(it.getBean(TransactionNameProvider::class.java)).isInstanceOf(SpringServletTransactionNameProvider::class.java)
             }
     }