+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 - Update normalization of metrics keys, tags and values ([#3332](https://github.com/getsentry/sentry-java/pull/3332))
 
+### Fixes
+
+- Add rate limit to Metrics ([#3334](https://github.com/getsentry/sentry-java/pull/3334))
+
 ## 7.7.0
 
 ### Features